extract smithy roll data
