+++ lib/index.ts
@@ -137,8 +137,9 @@ export * from './religion/Components/printInformation'
 export * from './religion/isInformationEmpty'
 
 export * from './smithy/_common'
-export * from './smithy/smithyData'
 export * from './smithy/smithyCustomers'
+export * from './smithy/smithyData'
+export * from './smithy/smithyRollData'
 
 export * from './src/arpa'
 export * from './src/articles'
